TYP: move stub-only helper types to pandas-stubs/_stubs_only
