+++ pandas-stubs/core/series.pyi
@@ -57,7 +57,27 @@ from pandas import (
     Timestamp,
 )
 from pandas._stubs_only import (
+    ArrayIndexSeriesTimedeltaNoSeq,
+    ArrayIndexTimedeltaNoSeq,
+    ElementOpsMixin,
+    NumListLike,
     OrderableT,
+    ScalarArrayIndexSeriesComplex,
+    ScalarArrayIndexSeriesJustComplex,
+    ScalarArrayIndexSeriesJustFloat,
+    ScalarArrayIndexSeriesJustInt,
+    ScalarArrayIndexSeriesReal,
+    ScalarArrayIndexSeriesTimedelta,
+    SeriesComplex,
+    SeriesReal,
+    Supports_ProtoAdd,
+    Supports_ProtoFloorDiv,
+    Supports_ProtoMul,
+    Supports_ProtoRAdd,
+    Supports_ProtoRFloorDiv,
+    Supports_ProtoRMul,
+    Supports_ProtoRTrueDiv,
+    Supports_ProtoTrueDiv,
     T_co,
     T_contra,
 )
@@ -78,27 +98,7 @@ from pandas.core.arrays.floating import FloatingArray
 from pandas.core.arrays.timedeltas import TimedeltaArray
 from pandas.core.base import (
     T_INTERVAL_NP,
-    ArrayIndexSeriesTimedeltaNoSeq,
-    ArrayIndexTimedeltaNoSeq,
-    ElementOpsMixin,
     IndexOpsMixin,
-    NumListLike,
-    ScalarArrayIndexSeriesComplex,
-    ScalarArrayIndexSeriesJustComplex,
-    ScalarArrayIndexSeriesJustFloat,
-    ScalarArrayIndexSeriesJustInt,
-    ScalarArrayIndexSeriesReal,
-    ScalarArrayIndexSeriesTimedelta,
-    SeriesComplex,
-    SeriesReal,
-    Supports_ProtoAdd,
-    Supports_ProtoFloorDiv,
-    Supports_ProtoMul,
-    Supports_ProtoRAdd,
-    Supports_ProtoRFloorDiv,
-    Supports_ProtoRMul,
-    Supports_ProtoRTrueDiv,
-    Supports_ProtoTrueDiv,
 )
 from pandas.core.frame import DataFrame
 from pandas.core.generic import NDFrame
